+++ lib/topics/weeks/scraping.dart
@@ -1,4 +1,3 @@
-
 import 'package:beautiful_soup_dart/beautiful_soup.dart';
 import 'package:intl/intl.dart';
 import 'package:lectio_wrapper/types/weeks/calendar_event.dart';
@@ -59,15 +58,24 @@ Future<Week> extractCalendar(BeautifulSoup soup, int year, int weekNum) async {
 }
 
 CalendarEvent extractModul(Bs4Element element) {
+  CalendarEventType type = CalendarEventType.regular;
   String status = "Uændret";
-  String id =
-      queriesFromSoup(element.getAttrValue('href') ?? "")['absid'] ?? "";
+  String? id = queriesFromSoup(element.getAttrValue('href') ?? "")['absid'];
+  if (id == null) {
+    type = CalendarEventType.test;
+    id = queriesFromSoup(element.getAttrValue('href') ?? "")['ProeveholdId'];
+  }
+  if (id == null) {
+    type = CalendarEventType.private;
+    id = queriesFromSoup(element.getAttrValue('href') ?? "")['aftaleid'];
+  }
   String title = "";
   DateTime start = DateTime.now();
   DateTime end = DateTime.now();
   String team = "";
   String room = "";
   String teacher = "";
+  String note = "";
   List<String> pieceInformation =
       element.getAttrValue('data-additionalinfo')!.split("\n");
   for (int j = 0; j < 2; j++) {
@@ -96,19 +104,26 @@ CalendarEvent extractModul(Bs4Element element) {
     } else if (states.contains(pieceInfo)) {
       status = pieceInfo;
     } else {
-      List<String> data = pieceInfo.split(": ");
+      List<String> data = pieceInfo.split(":");
       switch (data[0]) {
         case "Hold":
-          team = data[1];
+          team = data[1].trim();
           break;
         case "Lærer":
-          teacher = data[1];
+          teacher = data[1].trim();
           break;
         case "Lokale":
-          room = data[1];
+          room = data[1].trim();
+          break;
+        case "Note":
+          var index = pieceInformation.indexOf(pieceInfo);
+          if (index < pieceInformation.length - 1) {
+            note = pieceInformation.sublist(index + 1).join("\n");
+          }
           break;
       }
     }
   }
-  return CalendarEvent(status, title, team, teacher, room, id, start, end);
+  return CalendarEvent(status, title, team, teacher, room, id!, start, end,
+      type: type, note: note);
 }