+++ lib/types/weeks/calendar_event.dart
@@ -11,16 +11,21 @@ class Day {
   Day(this.informations, this.events, this.date);
 }
 
+enum CalendarEventType { test, regular, private }
+
 class CalendarEvent {
+  CalendarEventType type;
   String status;
   String title;
   String team;
   String teacher;
   String room;
   String id;
+  String note;
   DateTime start;
   DateTime end;
 
   CalendarEvent(this.status, this.title, this.team, this.teacher, this.room,
-      this.id, this.start, this.end);
+      this.id, this.start, this.end,
+      {this.type = CalendarEventType.regular, required this.note});
 }